% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocketClient.R, R/wssClient.R
\name{WebsocketClient}
\alias{WebsocketClient}
\alias{WebsocketClient}
\title{Create a websocket client}
\arguments{
\item{url}{The websocket URL.}

\item{onMessage}{A function called for each message received from the server.
Must take a single argument, the message string.}

\item{onDisconnected}{A function called with no arguments when either the client
or the server disconnect.}

\item{url}{The websocket URL.}

\item{onMessage}{A function called for each message received from the server.
Must take a single argument, the message string.}

\item{onDisconnected}{A function called with no arguments when either the client
or the server disconnect.}
}
\value{
a WebsocketClient instance

a WebsocketClient instance
}
\description{
Create a websocket client

Create a websocket client
}
\section{Usage}{

\preformatted{WebsocketClient$new(url, onMessage = identity, onDisconnected = identity)}


\preformatted{WebsocketClient$new(url, onMessage = identity, onDisconnected = identity)}
}

\examples{
# Create a websocket using the websocket.org test server
ws <- WebsocketClient$new("ws://echo.websocket.org/",
  onMessage = function(msg) {
    cat("Client got msg: ", msg, "\\n")
  },
  onDisconnected = function() {
    cat("Client disconnected\\n")
  }
)

# Send a message
ws$send("hello")

# Close the websocket after we're done with it
ws$close()
# Create a websocket using the websocket.org test server
ws <- WebsocketClient$new("ws://echo.websocket.org/",
  onMessage = function(msg) {
    cat("Client got msg: ", msg, "\\n")
  },
  onDisconnected = function() {
    cat("Client disconnected\\n")
  }
)

# Send a message
ws$send("hello")

# Close the websocket after we're done with it
ws$close()
}
