% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\name{wsOptions}
\alias{wsOptions}
\title{Manage WebSocket options}
\usage{
wsOptions(...)
}
\arguments{
\item{...}{Named options to be overridden.}
}
\value{
If \code{opts} is \code{NULL} (the default), returns the default options. Otherwise,
returns a named list that can be passed into \code{WebSocket$new()} as the \code{wsOptions}
parameter to customize the behavior of a WebSocket.
}
\description{
Lists the default options or assists in creating a custom configuration to
use when creating a new WebSocket.
}
\details{
The following options are available:
\describe{
  \item{\code{max_message_size}}{The maximum size of a message in bytes. If a message
  larger than this is sent, the connection will fail with the \code{message_too_big}
  protocol error.}
}
}
\examples{
ws <- WebSocket$new("ws://echo.websocket.org/", opts = wsOptions(max_message_size=1234))
}
