% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocketClient.R
\name{WebsocketClient}
\alias{WebsocketClient}
\title{Create a websocket client}
\arguments{
\item{url}{The websocket URL.}

\item{onMessage}{A function called for each message received from the server.
Must take a single argument, the message content. If the message is text,
the \code{onMessage} function will be passed a one-element character vector;
if the message is binary, the \code{onMessage} function will be passed a raw
vector.}

\item{onOpen}{A function called with no arguments when the connection is
established.}

\item{onClose}{A function called with no arguments when either the client or
the server disconnect.}

\item{onFail}{A function called with no arguments when the connection fails
while the handshake is bring processed.}

\item{headers}{A named list or character vector representing keys and values
of headers in the initial HTTP request.}
}
\value{
a WebsocketClient instance
}
\description{
Create a websocket client
}
\section{Usage}{

\preformatted{WebsocketClient$new(url, onMessage,
                     onOpen = function() {},
                     onClose = function() {},
                     onFail = function() {})
}
}

\examples{
# Create a websocket using the websocket.org test server
ws <- WebsocketClient$new("ws://echo.websocket.org/",
  onMessage = function(msg) {
    cat("Client got msg: ", msg, "\\n")
  },
  onClose = function() {
    cat("Client disconnected\\n")
  }
)

# Send a message
ws$send("hello")

# Close the websocket after we're done with it
ws$close()

}
